+++ app/Http/Controllers/EditInventoryController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use App\Genre;
+use App\Http\Requests\AddInventoryItemRequest;
+use App\Http\Requests\PatchInventoryItemRequest;
 use App\InventoryItem;
 use App\InventoryItemStatus;
 
@@ -19,5 +21,15 @@ class EditInventoryController extends Controller
         $genres = Genre::allNotTranslated();
         $inventoryStatuses = InventoryItemStatus::allTranslated();
         return view('edit-inventory', compact('inventoryItems', 'genres', 'inventoryStatuses'));
+    }
+
+    public function post(AddInventoryItemRequest $request)
+    {
+
+    }
+
+    public function patch(PatchInventoryItemRequest $request)
+    {
+
     }
 }