+++ resources/views/edit-inventory.blade.php
@@ -34,152 +34,171 @@
         @endslot
     @endheader
     <div class="container-fluid">
-
-            <div class="row">
-                <h2 class="ml-3">{{__('messages.edit_inventory.add_item')}}</h2>
-                <div class="col-md-12 table-responsive">
-                    <table class="table table-striped table-edit-inventory">
-                        <thead class="thead-light">
-                            <tr>
-                                <th colspan="1" class="border-right">{{__('Names')}}</th>
-                                <th colspan="3" class="border-right">{{__('Properties')}}</th>
-                                <th colspan="1">{{__('Add')}}</th>
-                            </tr>
-                            <tr>
-                                <th scope="col"></th>
-                                <th scope="col">{{__('Genres')}}</th>
-                                <th scope="col">{{__('Duration')}}</th>
-                                <th scope="col" class="border-right">{{__('Players')}}</th>
-                                <th scope="col"></th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            <tr>
-                                <td class="align-middle border-right">
+        <div class="row">
+            <h2 class="ml-3">{{__('messages.edit_inventory.add_item')}}</h2>
+            <div class="col-md-12 table-responsive">
+                <table class="table table-striped table-edit-inventory">
+                    <thead class="thead-light">
+                        <tr>
+                            <th colspan="1" class="border-right">{{__('Names')}}</th>
+                            <th colspan="3" class="border-right">{{__('Properties')}}</th>
+                            <th colspan="1">{{__('Add')}}</th>
+                        </tr>
+                        <tr>
+                            <th scope="col"></th>
+                            <th scope="col">{{__('Genres')}}</th>
+                            <th scope="col">{{__('Duration')}}</th>
+                            <th scope="col" class="border-right">{{__('Players')}}</th>
+                            <th scope="col"></th>
+                        </tr>
+                    </thead>
+                    <tbody>
+                        <tr>
+                            <td class="align-middle border-right">
+                                <div id="nameFr-field-new">
                                     <div class="input-group mb-1">
                                         <div class="input-group-prepend">
-                                            <label class="input-group-text" for="language-fr-new">
+                                            <label class="input-group-text" for="nameFr-new">
                                                 <i class="flag-icon flag-icon-fr flag-icon-squared"></i>
                                             </label>
                                         </div>
-                                        <input id="language-fr-new" name="language-fr-new" type="text" class="form-control" placeholder="{{__('French name')}}" autocomplete="off" form="add-game">
+                                        <input id="nameFr-new" name="nameFr" type="text" class="form-control" placeholder="{{__('French name')}}" autocomplete="off" form="add-item-form">
                                     </div>
+                                </div>
+                                <div id="nameEn-field-new">
                                     <div class="input-group mb-0">
                                         <div class="input-group-prepend">
-                                            <label class="input-group-text" for="language-gb-new">
+                                            <label class="input-group-text" for="nameEn-new">
                                                 <i class="flag-icon flag-icon-gb flag-icon-squared"></i>
                                             </label>
                                         </div>
-                                        <input id="language-gb-new" name="language-gb-new" type="text" class="form-control" placeholder="{{__('English name')}}" autocomplete="off" form="add-game">
+                                        <input id="nameEn-new" name="nameEn" type="text" class="form-control" placeholder="{{__('English name')}}" autocomplete="off" form="add-item-form">
                                     </div>
-                                </td>
-                                <td class="align-middle">
+                                </div>
+                            </td>
+                            <td class="align-middle">
+                                <div id="genres-field-new">
                                     <ul class="mb-0">
                                         <li class="plus-li">
                                             <a href="">{{__('messages.edit_inventory.add_new_genre')}}</a>
                                         </li>
                                     </ul>
-                                </td>
-                                <td class="align-middle">
+                                </div>
+                            </td>
+                            <td class="align-middle">
+                                <div id="durationMin-field-new">
                                     <div class="input-group mb-1">
                                         <div class="input-group-prepend">
-                                            <label class="input-group-text" for="duration-min-new">
+                                            <label class="input-group-text" for="durationMin-new">
                                                 {{__('Min')}}
                                             </label>
                                         </div>
-                                        <input id="duration-min-new" name="duration-min-new" type="number" min="0" class="form-control number-input" autocomplete="off" form="add-game">
-                                        <div class="input-group-append">
-                                            <label class="input-group-text" for="duration-min-new">
+                                        <input id="durationMin-new" name="durationMin" type="text" pattern="[0-9]*" class="form-control number-input" autocomplete="off" form="add-item-form">
+                                        <div class="input-group-append d-none d-xl-block">
+                                            <label class="input-group-text" for="durationMin-new">
                                                 {{strtolower(__('Minutes'))}}
                                             </label>
                                         </div>
                                     </div>
+                                </div>
+                                <div id="durationMax-field-new">
                                     <div class="input-group mb-0">
                                         <div class="input-group-prepend">
-                                            <label class="input-group-text" for="duration-max-new">
+                                            <label class="input-group-text" for="durationMax-new">
                                                 {{__('Max')}}
                                             </label>
                                         </div>
-                                        <input id="duration-max-new" name="duration-max-new" type="number" min="0" class="form-control number-input" autocomplete="off" form="add-game">
-                                        <div class="input-group-append">
-                                            <label class="input-group-text" for="duration-max-new">
+                                        <input id="durationMax-new" name="durationMax" type="text" pattern="[0-9]*" class="form-control number-input" autocomplete="off" form="add-item-form">
+                                        <div class="input-group-append d-none d-xl-block">
+                                            <label class="input-group-text" for="durationMax-new">
                                                 {{strtolower(__('Minutes'))}}
                                             </label>
                                         </div>
                                     </div>
-                                </td>
-                                <td class="align-middle border-right">
+                                </div>
+                            </td>
+                            <td class="align-middle border-right">
+                                <div id="playersMin-field-new">
                                     <div class="input-group mb-1">
                                         <div class="input-group-prepend">
-                                            <label class="input-group-text" for="players-min-new">
+                                            <label class="input-group-text" for="playersMin-new">
                                                 {{__('Min')}}
                                             </label>
                                         </div>
-                                        <input id="players-min-new" name="players-min-new" type="number" min="0" class="form-control number-input" autocomplete="off" form="add-game">
+                                        <input id="playersMin-new" name="playersMin" type="text" pattern="[0-9]*[1-9][0-9]*" class="form-control number-input" autocomplete="off" form="add-item-form">
                                     </div>
+                                </div>
+                                <div id="playersMax-field-new">
                                     <div class="input-group mb-0">
                                         <div class="input-group-prepend">
-                                            <label class="input-group-text" for="players-max-new">
+                                            <label class="input-group-text" for="playersMax-new">
                                                 {{__('Max')}}
                                             </label>
                                         </div>
-                                        <input id="players-max-new" name="players-max-new" type="number" min="0" class="form-control number-input" autocomplete="off" form="add-game">
+                                        <input id="playersMax-new" name="playersMax" type="text" pattern="[0-9]*[1-9][0-9]*" class="form-control number-input" autocomplete="off" form="add-item-form">
                                     </div>
-                                </td>
-                                <td class="align-middle">
-                                    <form id="add-game">
-                                        <button class="btn btn-sm btn-primary w-100">
-                                            {{__('Add')}}
-                                        </button>
-                                    </form>
-                                </td>
-                            </tr>
-                        </tbody>
-                    </table>
-                </div>
+                                </div>
+                            </td>
+                            <td class="align-middle">
+                                <form method="POST" action="/edit-inventory" id="add-item-form">
+                                    @csrf
+                                    <button class="btn btn-sm btn-primary w-100" id="add-item-submit-button" type="submit">
+                                        {{__('Add')}}
+                                    </button>
+                                </form>
+                            </td>
+                        </tr>
+                    </tbody>
+                </table>
             </div>
-            <div class="row">
-                <h2 class="ml-3">{{__('messages.edit_inventory.edit_items')}}</h2>
-                <div class="col-md-12 table-responsive">
-                    <table class="table table-striped table-edit-inventory">
-                        <thead class="thead-light">
-                            <tr>
-                                <th colspan="1" class="border-right">{{__('Names')}}</th>
-                                <th colspan="3" class="border-right">{{__('Properties')}}</th>
-                                <th colspan="3">{{__('Actions')}}</th>
-                            </tr>
-                            <tr>
-                                <th scope="col"></th>
-                                <th scope="col">{{__('Genres')}}</th>
-                                <th scope="col">{{__('Duration')}}</th>
-                                <th scope="col" class="border-right">{{__('Players')}}</th>
-                                <th scope="col">{{__('messages.edit_inventory.change_status')}}</th>
-                                <th scope="col">{{__('Delete')}}</th>
-                                <th scope="col">{{__('Confirm changes')}}</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            @foreach($inventoryItems as $inventoryItem)
-                                <tr id="modify-item-{{$inventoryItem->id}}">
-                                    <td class="align-middle border-right">
+        </div>
+        <div class="row">
+            <h2 class="ml-3">{{__('messages.edit_inventory.edit_items')}}</h2>
+            <div class="col-md-12 table-responsive">
+                <table class="table table-striped table-edit-inventory">
+                    <thead class="thead-light">
+                        <tr>
+                            <th colspan="1" class="border-right">{{__('Names')}}</th>
+                            <th colspan="3" class="border-right">{{__('Properties')}}</th>
+                            <th colspan="3">{{__('Actions')}}</th>
+                        </tr>
+                        <tr>
+                            <th scope="col"></th>
+                            <th scope="col">{{__('Genres')}}</th>
+                            <th scope="col">{{__('Duration')}}</th>
+                            <th scope="col" class="border-right">{{__('Players')}}</th>
+                            <th scope="col">{{__('messages.edit_inventory.change_status')}}</th>
+                            <th scope="col">{{__('Delete')}}</th>
+                            <th scope="col">{{__('Confirm changes')}}</th>
+                        </tr>
+                    </thead>
+                    <tbody>
+                        @foreach($inventoryItems as $inventoryItem)
+                            <tr id="modify-item-{{$inventoryItem->id}}">
+                                <td class="align-middle border-right">
+                                    <div id="nameFr-field-{{$inventoryItem->id}}">
                                         <div class="input-group mb-1">
                                             <div class="input-group-prepend">
-                                                <label class="input-group-text" for="language-fr-{{$inventoryItem->id}}">
+                                                <label class="input-group-text" for="nameFr-{{$inventoryItem->id}}">
                                                     <i class="flag-icon flag-icon-fr flag-icon-squared"></i>
                                                 </label>
                                             </div>
-                                            <input id="language-fr-{{$inventoryItem->id}}" name="language-fr-{{$inventoryItem->id}}" type="text" class="form-control" autocomplete="off" value="{{$inventoryItem->name_fr}}" form="edit-item-{{$inventoryItem->id}}">
+                                            <input id="nameFr-{{$inventoryItem->id}}" name="nameFr" type="text" class="form-control" placeholder="{{__('French name')}}" autocomplete="off" value="{{$inventoryItem->name_fr}}" form="edit-item-{{$inventoryItem->id}}-form">
                                         </div>
+                                    </div>
+                                    <div id="nameEn-field-{{$inventoryItem->id}}">
                                         <div class="input-group mb-0">
                                             <div class="input-group-prepend">
-                                                <label class="input-group-text" for="language-gb-{{$inventoryItem->id}}">
+                                                <label class="input-group-text" for="nameEn-{{$inventoryItem->id}}">
                                                     <i class="flag-icon flag-icon-gb flag-icon-squared"></i>
                                                 </label>
                                             </div>
-                                            <input id="language-gb-{{$inventoryItem->id}}" name="language-gb-{{$inventoryItem->id}}" type="text" class="form-control" autocomplete="off" value="{{$inventoryItem->name_en}}" form="edit-item-{{$inventoryItem->id}}">
+                                            <input id="nameEn-{{$inventoryItem->id}}" name="nameEn" type="text" class="form-control" placeholder="{{__('English name')}}" autocomplete="off" value="{{$inventoryItem->name_en}}" form="edit-item-{{$inventoryItem->id}}-form">
                                         </div>
-                                    </td>
-                                    <td class="align-middle">
+                                    </div>
+                                </td>
+                                <td class="align-middle">
+                                    <div id="genres-field-{{$inventoryItem->id}}">
                                         <ul class="mb-0">
                                             @foreach($inventoryItem->genres as $genre)
                                                 <li>
@@ -193,55 +212,68 @@
                                                 <a href="">{{__('messages.edit_inventory.add_new_genre')}}</a>
                                             </li>
                                         </ul>
-                                    </td>
-                                    <td class="align-middle">
+                                    </div>
+                                </td>
+                                <td class="align-middle">
+                                    <div id="durationMin-field-{{$inventoryItem->id}}">
                                         <div class="input-group mb-1">
                                             <div class="input-group-prepend">
-                                                <label class="input-group-text" for="duration-min-{{$inventoryItem->id}}">
+                                                <label class="input-group-text" for="durationMin-{{$inventoryItem->id}}">
                                                     {{__('Min')}}
                                                 </label>
                                             </div>
-                                            <input id="duration-min-{{$inventoryItem->id}}" name="duration-min-{{$inventoryItem->id}}" type="number" min="0" class="form-control number-input" autocomplete="off" value="{{$inventoryItem->duration->min}}" form="edit-item-{{$inventoryItem->id}}">
-                                            <div class="input-group-append">
-                                                <label class="input-group-text" for="duration-min-{{$inventoryItem->id}}">
+                                            <input id="durationMin-{{$inventoryItem->id}}" name="durationMin" type="text" pattern="[0-9]*" class="form-control number-input" autocomplete="off" value="{{$inventoryItem->duration->min}}" form="edit-item-{{$inventoryItem->id}}-form">
+                                            <div class="input-group-append d-none d-xl-block">
+                                                <label class="input-group-text" for="durationMin-{{$inventoryItem->id}}">
                                                     {{strtolower(__('Minutes'))}}
                                                 </label>
                                             </div>
                                         </div>
+                                    </div>
+                                    <div id="durationMax-field-{{$inventoryItem->id}}">
                                         <div class="input-group mb-0">
                                             <div class="input-group-prepend">
-                                                <label class="input-group-text" for="duration-max-{{$inventoryItem->id}}">
+                                                <label class="input-group-text" for="durationMax-{{$inventoryItem->id}}">
                                                     {{__('Max')}}
                                                 </label>
                                             </div>
-                                            <input id="duration-max-{{$inventoryItem->id}}" name="duration-max-{{$inventoryItem->id}}" type="number" min="0" class="form-control number-input" autocomplete="off" value="{{$inventoryItem->duration->max}}" form="edit-item-{{$inventoryItem->id}}">
-                                            <div class="input-group-append">
-                                                <label class="input-group-text" for="duration-max-{{$inventoryItem->id}}">
+                                            <input id="durationMax-{{$inventoryItem->id}}" name="durationMax" type="text" pattern="[0-9]*" class="form-control number-input" autocomplete="off" value="{{$inventoryItem->duration->max}}" form="edit-item-{{$inventoryItem->id}}-form">
+                                            <div class="input-group-append d-none d-xl-block">
+                                                <label class="input-group-text" for="durationMax-{{$inventoryItem->id}}">
                                                     {{strtolower(__('Minutes'))}}
                                                 </label>
                                             </div>
                                         </div>
-                                    </td>
-                                    <td class="align-middle border-right">
+                                    </div>
+                                </td>
+                                <td class="align-middle border-right">
+                                    <div id="playersMin-field-{{$inventoryItem->id}}">
                                         <div class="input-group mb-1">
                                             <div class="input-group-prepend">
-                                                <label class="input-group-text" for="players-min-{{$inventoryItem->id}}">
+                                                <label class="input-group-text" for="playersMin-{{$inventoryItem->id}}">
                                                     {{__('Min')}}
                                                 </label>
                                             </div>
-                                            <input id="players-min-{{$inventoryItem->id}}" name="players-min-{{$inventoryItem->id}}" type="number" min="0" class="form-control number-input" autocomplete="off" value="{{$inventoryItem->players->min}}" form="edit-item-{{$inventoryItem->id}}">
+                                            <input id="playersMin-{{$inventoryItem->id}}" name="playersMin" pattern="[0-9]*[1-9][0-9]*" class="form-control number-input" autocomplete="off" value="{{$inventoryItem->players->min}}" form="edit-item-{{$inventoryItem->id}}-form">
                                         </div>
+                                    </div>
+                                    <div id="playersMax-field-{{$inventoryItem}}">
                                         <div class="input-group mb-0">
                                             <div class="input-group-prepend">
-                                                <label class="input-group-text" for="players-max-{{$inventoryItem->id}}">
+                                                <label class="input-group-text" for="playersMax-{{$inventoryItem->id}}">
                                                     {{__('Max')}}
                                                 </label>
                                             </div>
-                                            <input id="players-max-{{$inventoryItem->id}}" name="players-max-{{$inventoryItem->id}}" type="number" min="0" class="form-control number-input" autocomplete="off" value="{{$inventoryItem->players->max}}" form="edit-item-{{$inventoryItem->id}}">
+                                            <input id="playersMax-{{$inventoryItem->id}}" name="playersMax" pattern="[0-9]*[1-9][0-9]*" class="form-control number-input" autocomplete="off" value="{{$inventoryItem->players->max}}" form="edit-item-{{$inventoryItem->id}}-form">
                                         </div>
-                                    </td>
-                                    <td class="align-middle">
-                                        <select id="status-{{$inventoryItem->id}}" name="status-{{$inventoryItem->id}}" class="custom-select form-control" autocomplete="off" form="edit-item-{{$inventoryItem->id}}">
+                                    </div>
+                                </td>
+                                <td class="align-middle">
+                                    <div id="status-field-{{$inventoryItem->id}}">
+                                        <label class="input-group-text d-none" for="status-{{$inventoryItem->id}}">
+                                            {{__('Status')}}
+                                        </label>
+                                        <select id="status-{{$inventoryItem->id}}" name="status" class="custom-select form-control" autocomplete="off" form="edit-item-{{$inventoryItem->id}}-form">
                                             @foreach($inventoryStatuses as $inventoryStatus)
                                                 <option value="{{$inventoryStatus->id}}"
                                                         @if($inventoryItem->status->id === $inventoryStatus->id)
@@ -256,26 +288,33 @@
                                                 </option>
                                             @endforeach
                                         </select>
-                                    </td>
-                                    <td class="align-middle"><button class="btn btn-sm btn-danger w-100">{{__('Delete')}}</button></td>
-                                    <td class="align-middle">
-                                        <form id="edit-item-{{$inventoryItem->id}}">
-                                            <button class="btn btn-sm btn-primary w-100">
-                                                {{__('Confirm')}}
-                                            </button>
-                                        </form>
-                                    </td>
-                                </tr>
-                            @endforeach
-                        </tbody>
-                    </table>
-                </div>
+                                    </div>
+                                </td>
+                                <td class="align-middle">
+                                    <button class="btn btn-sm btn-danger w-100">
+                                        {{__('Delete')}}
+                                    </button>
+                                </td>
+                                <td class="align-middle">
+                                    <form id="edit-item-{{$inventoryItem->id}}-form">
+                                        @csrf
+                                        <button class="btn btn-sm btn-primary w-100" id="edit-item-{{$inventoryItem->id}}-submit-button" type="submit">
+                                            {{__('Confirm')}}
+                                        </button>
+                                    </form>
+                                </td>
+                            </tr>
+                        @endforeach
+                    </tbody>
+                </table>
             </div>
-
+        </div>
     </div>
-    <script type="javascript">
-        $(document).ready(function() {
-            localStorage.clear();
-        });
-    </script>
 @endsection
+
+@push('scripts')
+    <script type="text/javascript">
+        const inventoryItems = {!! json_encode($inventoryItems)!!};
+    </script>
+    <script type="text/javascript" src="{{URL::asset('js/editInventory.js')}}"></script>
+@endpush